Adds Netlify deployment configuration.
This deploys an empty directory with a Netlify configuration to always redirect directly to `blog.dwac.dev`.

Also added some documentation about how to do the deployment and linking to Netlify's configuration settings.

For now, deployments are manual because this should not change frequently. We can always revisit this in the future as the site grows.